+++ src/traps.cpp
@@ -7,29 +7,29 @@
     Class Trap.
 
     A single trap species.
-    
+    
     Controls the density of traps and the timescales/probabilities of
     capture and release, along with utilities for the watermarking tracking
     of trap states and the calculation of capture and release.
 
     Parameters
     ----------
-    density : float
+    density : double
         The density of the trap species in a pixel.
-        
-    release_timescale : float
+        
+    release_timescale : double
         The release timescale of the trap, in the same units as the time
         spent in each pixel or phase.
-        
-    capture_timescale : float
+        
+    capture_timescale : double
         The capture timescale of the trap. Default 0 for instant capture.
-        
+        
     Attributes
     ----------
-    capture_rate, emission_rate : float
+    capture_rate, emission_rate : double
         The capture and emission rates (Lindegren (1998) section 3.2).
 */
-Trap::Trap(float density, float release_timescale, float capture_timescale)
+Trap::Trap(double density, double release_timescale, double capture_timescale)
     : density(density),
       release_timescale(release_timescale),
       capture_timescale(capture_timescale) {
@@ -42,41 +42,41 @@ Trap::Trap(float density, float release_timescale, float capture_timescale)
 }
 
 /*
-       Calculate the fraction of filled traps after an amount of elapsed time.
+    Calculate the fraction of filled traps after an amount of elapsed time.
 
     Parameters
     ----------
-    time_elapsed : float
+    time_elapsed : double
         The total time elapsed since the traps were filled, in the same
         units as the trap timescales.
 
     Returns
     -------
-    fill_fraction : float
+    fill_fraction : double
         The remaining fraction of filled traps.
 */
-float Trap::fill_fraction_from_time_elapsed(float time_elapsed) {
+double Trap::fill_fraction_from_time_elapsed(double time_elapsed) {
     return exp(-time_elapsed / release_timescale);
 }
 
 /*
     Class TrapInstantCapture.
-    
+    
     For the old release-then-instant-capture algorithm.
 
     Parameters
     ----------
-    density : float
+    density : double
         The density of the trap species in a pixel.
-        
-    release_timescale : float
+        
+    release_timescale : double
         The release timescale of the trap, in the same units as the time
         spent in each pixel or phase.
-        
+        
     Attributes
     ----------
-    emission_rate : float
+    emission_rate : double
         The emission rate (Lindegren (1998) section 3.2).
 */
-TrapInstantCapture::TrapInstantCapture(float density, float release_timescale)
+TrapInstantCapture::TrapInstantCapture(double density, double release_timescale)
     : Trap(density, release_timescale, 0.0) {}
